requirements.txt
aiohttp==2.3.7
aiomysql==0.0.11
async-timeout==2.0.0
asyncio==3.4.3
chardet==3.0.4
configparser==3.5.0
idna==2.6
motor==1.2.0
multidict==3.3.2
mysqlclient==1.3.12
netifaces==0.10.6
pymongo==3.6.0
PyMySQL==0.8.0
pytz==2017.3
PyYAML==3.12
six==1.11.0
trafaret==1.0.2
trafaret-config==1.0.1
yarl==0.17.0
hashids==1.2.0
PyJwt